literals package
