+++ src/main/java/ch/njol/skript/literals/LitDoubleMaxValue.java
@@ -1,7 +1,10 @@
-package ch.njol.skript.test.runner;
+package ch.njol.skript.literals;
 
 import ch.njol.skript.Skript;
-import ch.njol.skript.doc.NoDoc;
+import ch.njol.skript.doc.Description;
+import ch.njol.skript.doc.Example;
+import ch.njol.skript.doc.Name;
+import ch.njol.skript.doc.Since;
 import ch.njol.skript.lang.Expression;
 import ch.njol.skript.lang.ExpressionType;
 import ch.njol.skript.lang.SkriptParser.ParseResult;
@@ -10,13 +13,14 @@ import ch.njol.util.Kleenean;
 import org.bukkit.event.Event;
 import org.jetbrains.annotations.Nullable;
 
-@NoDoc
+@Name("Maximum Double Value")
+@Description("A number representing the maximum value of a double number type.")
+@Example("if {_number} >= maximum double value:")
+@Since("INSERT VERSION")
 public class LitDoubleMaxValue extends SimpleLiteral<Double> {
 
 	static {
-		if (TestMode.ENABLED) {
-			Skript.registerExpression(LitDoubleMaxValue.class, Double.class, ExpressionType.SIMPLE, "double max[imum] value");
-		}
+		Skript.registerExpression(LitDoubleMaxValue.class, Double.class, ExpressionType.SIMPLE, "max[imum] double value");
 	}
 
 	public LitDoubleMaxValue() {
@@ -30,7 +34,7 @@ public class LitDoubleMaxValue extends SimpleLiteral<Double> {
 
 	@Override
 	public String toString(@Nullable Event event, boolean debug) {
-		return "double max value";
+		return "max double value";
 	}
 
 }
