+++ src/main/java/ch/njol/skript/literals/LitFloatMinValue.java
@@ -1,7 +1,10 @@
-package ch.njol.skript.test.runner;
+package ch.njol.skript.literals;
 
 import ch.njol.skript.Skript;
-import ch.njol.skript.doc.NoDoc;
+import ch.njol.skript.doc.Description;
+import ch.njol.skript.doc.Example;
+import ch.njol.skript.doc.Name;
+import ch.njol.skript.doc.Since;
 import ch.njol.skript.lang.Expression;
 import ch.njol.skript.lang.ExpressionType;
 import ch.njol.skript.lang.SkriptParser.ParseResult;
@@ -10,13 +13,14 @@ import ch.njol.util.Kleenean;
 import org.bukkit.event.Event;
 import org.jetbrains.annotations.Nullable;
 
-@NoDoc
+@Name("Minimum Float Value")
+@Description("A number representing the minimum value of a float number type.")
+@Example("if {_number} <= minimum float value:")
+@Since("INSERT VERSION")
 public class LitFloatMinValue extends SimpleLiteral<Float> {
 
 	static {
-		if (TestMode.ENABLED) {
-			Skript.registerExpression(LitFloatMinValue.class, Float.class, ExpressionType.SIMPLE, "float min[imum] value");
-		}
+		Skript.registerExpression(LitFloatMinValue.class, Float.class, ExpressionType.SIMPLE, "min[imum] float value");
 	}
 
 	public LitFloatMinValue() {
@@ -30,7 +34,7 @@ public class LitFloatMinValue extends SimpleLiteral<Float> {
 
 	@Override
 	public String toString(@Nullable Event event, boolean debug) {
-		return "float min value";
+		return "min float value";
 	}
 
 }
