+++ src/test/skript/tests/regressions/7645-floating point timespan math.sk
@@ -9,4 +9,4 @@ test "floating point timespan math":
 	assert (5 seconds / 2.5) is 2 seconds with "failed floating point division"
 	assert (5 seconds / infinity value) is 0 seconds with "Division by infinity didn't return 0 seconds"
 
-	assert (5 seconds * 10 ^ 308) is {@max-timespan} with "failed to clamp to the long max value"
+	assert (5 seconds * 10 ^ 308) is {@max-timespan} with "failed to clamp to the max long value"
