+++ src/test/skript/tests/syntaxes/expressions/ExprRemainingAir.sk
@@ -13,6 +13,6 @@ test "remaining air":
 	assert remaining air of {_e} is 0 seconds with "removing did not limit to 0 seconds"
 
 	add 10 years to remaining air of {_e}
-	assert ticks of remaining air of {_e} is integer max value with "adding did not limit to max int value"
+	assert ticks of remaining air of {_e} is max integer value with "adding did not limit to max int value"
 
 	delete the entity within {_e}
